Update inferCNV.R
Fix to allow floats to be used as cutoff instead of only integers.

Former-commit-id: ace7d1a00bfe2bc94439324d8539e84755a80e25

+++ scripts/inferCNV.R
@@ -155,7 +155,7 @@ pargs <- optparse::add_option(pargs, c("--contig_lab_size"),
                                          "[Default %default]"))
 
 pargs <- optparse::add_option(pargs, c("--cutoff"),
-                              type="integer",
+                              type="numeric",
                               default=0,
                               action="store",
                               dest="cutoff",
